game view
Created display of game in grid view. Depending on status of the game user will be able to perform a move on board. In case game is complete it will show the winner or if it's a draw. In case user is viewer he won't be able to perform any actions on the view. In case status is set to  user will be able to join if he didn't create the game.